+++ Craft.Net.Server/Worlds/Generation/IWorldGenerator.cs
@@ -5,7 +5,7 @@ namespace Craft.Net.Server.Worlds.Generation
     public interface IWorldGenerator
     {
         string LevelType { get; }
-        double Seed { get; set; }
+        long Seed { get; set; }
         Vector3 SpawnPoint { get; }
         /// <summary>
         /// Generates one chunk at the given position.
